# backend/internal/handlers/task_handler.go
package handlers

import (
	"encoding/json"
	"errors"
	"log/slog"
	"net/http"
	"strconv"
	"strings"

	"github.com/probablynotvaish/task-management-system/backend/internal/middleware"
	"github.com/probablynotvaish/task-management-system/backend/internal/models"
	"github.com/probablynotvaish/task-management-system/backend/internal/repository"
	"github.com/probablynotvaish/task-management-system/backend/internal/service"
	"go.mongodb.org/mongo-driver/v2/bson"
)

type TaskHandler struct {
	service *service.TaskService
}

func NewTaskHandler(service *service.TaskService) *TaskHandler {
	return &TaskHandler{service: service}
}

// GetTasks godoc
// @Summary List user tasks
// @Description Retrieve a list of tasks for the authenticated user, supporting filtering, sorting, pagination, and search.
// @Tags Tasks
// @Accept json
// @Produce json
// @Param status query string false "Filter tasks by status (e.g. to_do, in_progress, completed, archived)"
// @Param priority query string false "Filter tasks by priority (low, medium, high)"
// @Param search query string false "Search query to match task title/description"
// @Param page query integer false "Page number"
// @Param page_size query integer false "Number of items per page"
// @Param sort_by query string false "Field to sort by (e.g. due_date, priority, created_at)"
// @Param sort_dir query string false "Sorting direction: asc or desc"
// @Success 200 {object} models.PaginatedResponse
// @Failure 401 {object} map[string]string "error: unauthorized"
// @Failure 500 {object} map[string]string "error: failed to fetch tasks"
// @Security Bearer
// @Router /api/tasks [get]
func (h *TaskHandler) GetTasks(w http.ResponseWriter, r *http.Request) {
	userID, ok := middleware.GetUserID(r.Context())
	if !ok {
		writeJSON(w, http.StatusUnauthorized, map[string]string{"error": "unauthorized"})
		return
	}

	query := r.URL.Query()

	page, _ := strconv.Atoi(query.Get("page"))
	pageSize, _ := strconv.Atoi(query.Get("page_size"))

	sortDir := -1
	if query.Get("sort_dir") == "asc" {
		sortDir = 1
	}

	filter := models.TaskFilter{
		Status:   models.TaskStatus(query.Get("status")),
		Priority: models.TaskPriority(query.Get("priority")),
		Search:   strings.TrimSpace(query.Get("search")),
		Page:     page,
		PageSize: pageSize,
		SortBy:   query.Get("sort_by"),
		SortDir:  sortDir,
	}

	result, err := h.service.ListTasks(r.Context(), userID, filter)
	if err != nil {
		slog.Error("failed to list tasks", "error", err)
		writeJSON(w, http.StatusInternalServerError, map[string]string{"error": "failed to fetch tasks"})
		return
	}

	writeJSON(w, http.StatusOK, result)
}

// CreateTask godoc
// @Summary Create a new task
// @Description Create a new task for the authenticated user.
// @Tags Tasks
// @Accept json
// @Produce json
// @Param request body models.TaskDTO true "Task DTO body"
// @Success 201 {object} models.Task
// @Failure 400 {object} map[string]string "error: invalid request body or create task failed"
// @Failure 401 {object} map[string]string "error: unauthorized"
// @Security Bearer
// @Router /api/tasks [post]
func (h *TaskHandler) CreateTask(w http.ResponseWriter, r *http.Request) {
	userID, ok := middleware.GetUserID(r.Context())
	if !ok {
		writeJSON(w, http.StatusUnauthorized, map[string]string{"error": "unauthorized"})
		return
	}

	var payload models.TaskDTO
	if err := json.NewDecoder(r.Body).Decode(&payload); err != nil {
		writeJSON(w, http.StatusBadRequest, map[string]string{"error": "invalid request body"})
		return
	}

	task, err := h.service.CreateTask(r.Context(), userID, payload)
	if err != nil {
		slog.Warn("create task failed", "error", err)
		writeJSON(w, http.StatusBadRequest, map[string]string{"error": err.Error()})
		return
	}

	writeJSON(w, http.StatusCreated, task)
}

// UpdateTask godoc
// @Summary Update an existing task
// @Description Update an existing task's fields for the authenticated user by ID.
// @Tags Tasks
// @Accept json
// @Produce json
// @Param id path string true "Task ID"
// @Param request body models.Task true "Task fields to update"
// @Success 200 {object} models.Task
// @Failure 400 {object} map[string]string "error: invalid request body or ID format"
// @Failure 401 {object} map[string]string "error: unauthorized"
// @Failure 404 {object} map[string]string "error: task not found"
// @Failure 500 {object} map[string]string "error: failed to update task"
// @Security Bearer
// @Router /api/tasks/{id} [patch]
func (h *TaskHandler) UpdateTask(w http.ResponseWriter, r *http.Request) {
	userID, ok := middleware.GetUserID(r.Context())
	if !ok {
		writeJSON(w, http.StatusUnauthorized,
			map[string]string{"error": "unauthorized"})
		return
	}

	id := r.PathValue("id")

	var task models.Task

	if err := json.NewDecoder(r.Body).Decode(&task); err != nil {
		writeJSON(w, http.StatusBadRequest,
			map[string]string{"error": "invalid request body"})
		return
	}

	objectID, err := bson.ObjectIDFromHex(id)
	if err != nil {
		writeJSON(w, http.StatusBadRequest,
			map[string]string{"error": "invalid ID format"})
		return
	}

	task.ID = objectID

	if err := h.service.UpdateTask(r.Context(), userID, &task); err != nil {
		switch {
		case errors.Is(err, service.ErrTaskIDRequired):
			writeJSON(w, http.StatusBadRequest, map[string]string{"error": "task ID is required"})
			return
		case errors.Is(err, repository.ErrTaskNotFound):
			writeJSON(w, http.StatusNotFound, map[string]string{"error": "task not found"})
			return
		default:
			slog.Error("update task failed", "error", err)
			writeJSON(w, http.StatusInternalServerError, map[string]string{"error": "failed to update task"})
			return
		}
	}

	writeJSON(w, http.StatusOK, task)
}

// DeleteTask godoc
// @Summary Delete a task
// @Description Delete a task by ID for the authenticated user.
// @Tags Tasks
// @Produce json
// @Param id path string true "Task ID"
// @Success 200 {object} map[string]string "message: task deleted successfully"
// @Failure 400 {object} map[string]string "error: invalid ID format or ID is required"
// @Failure 401 {object} map[string]string "error: unauthorized"
// @Failure 404 {object} map[string]string "error: task not found"
// @Failure 500 {object} map[string]string "error: failed to delete task"
// @Security Bearer
// @Router /api/tasks/{id} [delete]
func (h *TaskHandler) DeleteTask(w http.ResponseWriter, r *http.Request) {
	userID, ok := middleware.GetUserID(r.Context())
	if !ok {
		writeJSON(w, http.StatusUnauthorized, map[string]string{"error": "unauthorized"})
		return
	}

	id := r.PathValue("id")
	if id == "" {
		writeJSON(w, http.StatusBadRequest, map[string]string{"error": "task ID is required"})
		return
	}

	objectID, err := bson.ObjectIDFromHex(id)
	if err != nil {
		writeJSON(w, http.StatusBadRequest, map[string]string{"error": "invalid ID format"})
		return
	}

	if err := h.service.DeleteTask(r.Context(), userID, objectID); err != nil {
		if errors.Is(err, repository.ErrTaskNotFound) {
			writeJSON(w, http.StatusNotFound, map[string]string{"error": "task not found"})
			return
		}

		slog.Error("delete task failed", "error", err)

		writeJSON(w, http.StatusInternalServerError,
			map[string]string{"error": "failed to delete task"})
		return
	}

	writeJSON(w, http.StatusOK,
		map[string]string{"message": "task deleted successfully"})
}